Move core stuff from controller to service provider. #206
